\name{mlogitBMA-internal}
\Rdversion{1.1}
\alias{bic.glm.bg}
\alias{bic.glm.bg.data.frame}
\alias{bic.glm.bg.formula}
\alias{bic.glm.bg.matrix}
\alias{do.estimate.mlogit}
\alias{do.mlogit2logit}
\alias{get.mnl.data}
\alias{mnl.loglikelihood}
\alias{mnl.probabilities}
\alias{mnl.spec.reduce}
\alias{mnl.utilities}
\alias{print.summary.mnl}
\title{
Internal Functions of mlogitBMA
}

\description{
These functions are not to be used directly by the user.
}

\keyword{internal}


